Move Plugin interface to dedicated interface include directory and add interface library for client tools to link against
