+++ tools/cage/include/cage/interface/CaGePlugin.h
